new more awesome flatten function
with tests
