+++ project_generator/util.py
@@ -23,9 +23,27 @@ def uniqify(_list):
     # see: http://stackoverflow.com/questions/480214/how-do-you-remove-duplicates-from-a-list-in-python-whilst-preserving-order/29898968#29898968
     return reduce(lambda r, v: v in r[1] and r or (r[0].append(v) or r[1].add(v)) or r, _list, ([], set()))[0]
 
-def flatten_list(_list):
-    all_items = [item if type(item) is not list else sublist for sublist in _list for item in sublist]
-    return uniqify(all_items)
+def merge_recursive(*args):
+    if all(isinstance(x, dict) for x in args):
+        output = {}
+        keys = reduce(operator.or_, [set(x) for x in args])
+
+        for key in keys:
+            # merge all of the ones that have them
+            output[key] = merge_recursive(*[x[key] for x in args if key in x])
+
+        return output
+    else:
+        return reduce(operator.add, args)
+
+
+def flatten(*args):
+    for x in args:
+        if hasattr(x, '__iter__'):
+            for y in flatten(*x):
+                yield y
+        else:
+            yield x
 
 def unicode_available():
     return locale.getdefaultlocale()[1] == 'UTF-8'
